% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.mean.R
\name{est.density.minq2max}
\alias{est.density.minq2max}
\title{Estimating Unknown Parameters using Minimum, Median and Maximum}
\usage{
est.density.minq2max(
   min = NULL, 
   med = NULL, 
   max = NULL, 
   n = NULL, 
   opt = TRUE
   )
}
\arguments{
\item{min}{numeric value representing the sample minimum.}

\item{med}{numeric value representing the median of the sample.}

\item{max}{numeric value representing the sample maximum.}

\item{n}{numeric value specifying the sample size.}

\item{opt}{logical value indicating whether to apply the optimisation step in estimating parameters using theoretical quantiles.
The default value is \code{TRUE}.}
}
\value{
\itemize{
\item \code{parameters}: named numeric vector representing the estimated parameters ('location', 'scale', 'skewing') of SLD.
\item \code{mean}: numeric value of the estimated mean of the sample using SLD.
\item \code{sd}: numeric value of the estimated standard deviation of the sample using SLD.
}
}
\description{
This function provide estimates for the parameters of skew logistic distribution (SLD), the sample mean and the standard deviation using 3-number summary \{minimum, median (\eqn{q_2}), maximum\} from a study with sample size \eqn{n},
using the method explained in De Livera et al. (2024).
}
\details{
De Livera et al., (2024) proposed using the skew logistic distribution (SLD) to estimate unknown parameters for studies reporting 3-number summaries in the meta-analysis context.

The quantile-based skew logistic distribution, introduced by Gilchrist (2000) and further modified by van Staden and King (2015)
is used to approximate the sample's distribution using 3-point summaries.
The SLD quantile function is defined using three parameters: a location parameter (\eqn{\lambda}), a scale parameter (\eqn{\eta}), and a skewing parameter (\eqn{\delta}).
The parameters of the SLD are estimated by formulating and solving a set of simultaneous equations which relate the estimated sample quantiles to their theoretical counterparts of the SLD.
}
\examples{
#Generate 3-number summary data
set.seed(123)
n <- 1000
x <- stats::rlnorm(n, 4, 0.3)
quants <- c(min(x), stats::quantile(x, probs = 0.5), max(x))

#Estimate SLD parameters using 3-number summary
params <- est.density.minq2max(min = quants[1], med = quants[2], max = quants[3], 
                               n=n, opt=TRUE)$parameters
params

}
\references{
Alysha De Livera, Luke Prendergast, and Udara Kumaranathunga. A novel density-based approach for estimating unknown means, distribution visualisations, and meta-analyses of quantiles, 2024. Pre-print available here: \url{https://arxiv.org/abs/2411.10971}.

Warren Gilchrist. Statistical modelling with quantile functions. Chapman and Hall/CRC, 2000.

P. J. van Staden and R. A. R. King. The quantile-based skew logistic distribution.  \emph{Statistics & Probability Letters}, 96:109–116, 2015.

R. King and P. van Staden. sld: Estimation and Use of the Quantile-Based Skew Logistic Distribution. R package version 1.0.1, 2022. \doi{10.32614/CRAN.package.sld}. \url{https://CRAN.R-project.org/package=sld}.
}
\seealso{
\code{\link[=est.density.five]{est.density.five()}}, \code{\link[=est.density.q1q2q3]{est.density.q1q2q3()}}
}
